hawk/requirements.txt
celery==3.1.19
django==1.8.7
djangorestframework==3.3.1
django_filter==0.11.0
django-oauth-toolkit==0.9.0
gunicorn==19.4.1
markdown==2.6.5
psycopg2==2.6.1
pygments==2.0.2
pillow==3.0.0
simplejson==3.8.1
yolk==0.4.3
tornado==4.3
pika==0.10.0
tweepy==3.5.0
scapy==2.3.1
pcapy==0.10.9
motor==0.5
sockjs-tornado==1.0.2
